[Bug fixes] pqeqcoeff was defined integer insteads of real(8). Changed it to real(8) - Cleanup comm_sc.F90 routine
 Changes to be committed:
	modified:   src/comm_sc.F90
	modified:   src/init.F90
	modified:   src/module.F90

+++ src/module.F90
@@ -217,7 +217,7 @@ integer,allocatable :: nbplist(:,:)
 integer,allocatable :: nbplist_sc(:,:)
 
 !<pqeqcoeff> storing frequently reused PQEq coefficients in get_hsh_sc to save redundant computation 
-integer,allocatable :: pqeqcoeff(:,:,:)
+real(8),allocatable :: pqeqcoeff(:,:,:)
 
 real(8),allocatable :: fpqeq(:) !fpqeq become global variable for SC algorithm
